+++ _examples/allowlist/plugin.go
@@ -20,7 +20,7 @@ type plugin struct {
 	next corednsplugin.Handler
 }
 
-func NewPlugin(next corednsplugin.Handler) corednsplugin.Handler {
+func NewPlugin(next corednsplugin.Handler, rawOpts string) corednsplugin.Handler {
 	return plugin{next: next}
 }